@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { Route, Switch } from 'react-router-dom';
+import { Route, Routes, BrowserRouter } from 'react-router-dom';
 
 import Layout from './hoc/Layout/Layout'
 import LearningUnitBox from './components/LearningUnitBox/LearningUnitBox'
@@ -7,14 +7,16 @@ import ExpressionList from './containers/Expressions/ExpressionList'
 
 class App extends Component {
   render() {
-    return (
-      <Layout>
-        <Switch>
-          <Route path="/learn" component={LearningUnitBox} />
-          <Route path="/review" component={ExpressionList} />
-          <Route path="/" exact component={ExpressionList} />
-        </Switch>
-      </Layout>
+    return (      
+      <BrowserRouter>
+        <Layout>          
+          <Routes>
+            <Route path="/learn" element={<LearningUnitBox />} />
+            <Route path="/review" element={<ExpressionList />} />
+            <Route path="/" exact element={<ExpressionList />} />
+          </Routes>          
+        </Layout>      
+      </BrowserRouter>          
     )
   }
 }
