@@ -3,16 +3,16 @@ import React from 'react';
 import { useStore } from '../../hooks-store/store';
 import axios from '../../axios-wordyapp';
 
-import ListItem from '@material-ui/core/ListItem';
-import ListItemAvatar from '@material-ui/core/ListItemAvatar';
-import ListItemSecondaryAction from '@material-ui/core/ListItemSecondaryAction';
-import ListItemText from '@material-ui/core/ListItemText';
-import Avatar from '@material-ui/core/Avatar';
-import IconButton from '@material-ui/core/IconButton';
-import LabelIcon from '@material-ui/icons/Label';
-import DeleteIcon from '@material-ui/icons/Delete';
-import MoreVert from '@material-ui/icons/MoreVert';
-import Tooltip from '@material-ui/core/Tooltip';
+import ListItem from '@mui/material/ListItem';
+import ListItemAvatar from '@mui/material/ListItemAvatar';
+import ListItemSecondaryAction from '@mui/material/ListItemSecondaryAction';
+import ListItemText from '@mui/material/ListItemText';
+import Avatar from '@mui/material/Avatar';
+import IconButton from '@mui/material/IconButton';
+import LabelIcon from '@mui/icons-material/Label';
+import DeleteIcon from '@mui/icons-material/Delete';
+import MoreVert from '@mui/icons-material/MoreVert';
+import Tooltip from '@mui/material/Tooltip';
 
 import ExpressionIsLearning from './ExpressionIsLearning';
 
