@@ -1,12 +1,12 @@
 import React from 'react';
 
-import { makeStyles } from '@material-ui/core/styles';
-import InputLabel from '@material-ui/core/InputLabel';
-import FormControl from '@material-ui/core/FormControl';
-import Select from '@material-ui/core/Select';
-import MenuItem from '@material-ui/core/MenuItem';
+//import { makeStyles } from '@mui/styles/makeStyles';
+import InputLabel from '@mui/material/InputLabel';
+import FormControl from '@mui/material/FormControl';
+import Select from '@mui/material/Select';
+import MenuItem from '@mui/material/MenuItem';
 
-const useStyles = makeStyles((theme) => ({
+/* const useStyles = makeStyles((theme) => ({
   formControl: {
     minWidth: 420,
     marginBottom: theme.spacing(2),
@@ -15,16 +15,16 @@ const useStyles = makeStyles((theme) => ({
     marginTop: theme.spacing(2),
   },
 }));
-
+ */
 const LearuningUnitSelect = (props) => {
-  const classes = useStyles();
+  //const classes = useStyles();
 
-  const handleUnitSelectChange = (e) => {
+  const handleUnitSelectChange = (e) => {    
     props.onLearningUnitChange(e.target.value);
   };
 
   return (
-    <FormControl variant="outlined" className={classes.formControl}>
+    <FormControl variant="outlined">
       <InputLabel htmlFor="outlined-lu-native-simple">Learning Unit</InputLabel>
       <Select
         onChange={handleUnitSelectChange}
