@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Typography from '@material-ui/core/Typography';
+import Typography from '@mui/material/Typography';
 
 const learningExpression2 = (props) => (
   <Typography variant="subtitle1" color="textSecondary">
