@@ -1,25 +1,23 @@
 import React from 'react';
 import LearningUnitSelect from './LearningUnitSelect'
 
-class LearningUnitHeader extends React.Component {
-  constructor(props) {
+const LearningUnitHeader = (props) => {
+  /* constructor(props) {
     super(props);
     this.handleLearuningUnitChange = this.handleLearuningUnitChange.bind(this);
-  }
+  } */
 
-  handleLearuningUnitChange(learningUnit) {
-    this.props.onLearningUnitChange(learningUnit);
-  }
-
-  render() {
-    return (
-      <div>
-        <LearningUnitSelect
-          units={this.props.units}
-          onLearningUnitChange={this.handleLearuningUnitChange} />
-      </div>
-    );
-  }
+  const handleLearuningUnitChange = (learningUnit) => {    
+    props.onLearningUnitChange(learningUnit);
+  };
+  
+  return (
+    <div>
+      <LearningUnitSelect
+        units={props.units}
+        onLearningUnitChange={handleLearuningUnitChange} />
+    </div>
+  );  
 }
 
 export default LearningUnitHeader;
