@@ -1,6 +1,5 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
-import { BrowserRouter } from 'react-router-dom';
+import ReactDOM from 'react-dom/client';
 
 import './index.css';
 import configureExpressionsStore from './hooks-store/expressions-store';
@@ -11,13 +10,20 @@ import App from './App';
 (async () => {await configureLUStore()})();
 configureExpressionsStore();
 
-ReactDOM.render(
+/* ReactDOM.render(
   <React.StrictMode>
     <BrowserRouter>
       <App />
     </BrowserRouter>
   </React.StrictMode>,
   document.getElementById('root')
+); */
+
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
+  <React.StrictMode>    
+      <App />    
+  </React.StrictMode>
 );
 
 // If you want to start measuring performance in your app, pass a function
